+++ TrainingAnalyticsPlatform/ingestion/constants.py
@@ -4,5 +4,5 @@
 METADATA_SCHEMA_VERSION = "1.0.0"
 LAPS_SCHEMA_VERSION = "1.0.0"
 FIT_ANALYSIS_VERSION = "v1.0.0"
-INGEST_VERSION = "v15.4.0"  # (v15.3.0->v15.4.0) MINOR: pre-download Garmin manufacturer pre-filter using cached list metadata
+INGEST_VERSION = "v15.5.0"  # (v15.4.0->v15.5.0) MINOR: promote Garmin activity-list training/cycling/environmental fields into canonical enrichment metadata